requirements.txt
# Kumbuka dependencies
# Install with: pip install -r requirements.txt

numpy>=1.24.0
sounddevice>=0.4.6
httpx>=0.24.0
pyobjc-framework-CoreAudio>=10.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
